Add headers for block and color.
Color is a simple struct to hold RGB values and may not actually be
needed.

Block, on the other hand, is a start for board logic and represents a
physical block on the board. It has an x, y, and color value.

Cell is another type that just represents a cell on the board that
potentially contains a block. Right now it's a simple typedef for
std::optional<Block>, we can expand this into a deeper class if
necessary.