Add tetromino.hpp, modify block.hpp
In considering the different kinds of designs that might work best for
this project, I am considering an "ownership model". This works like so:

- The gameboard is a matrix 10x20 matrix of "Blocks."

- These Blocks can be either active, or inactive.
    - If they are active, they have a color and are lit up.
    - If they are inactive, they do not have a color, and are not
      visible.

- These Blocks may be owned by a Tetromino. A Tetromino must own exactly
  4 Blocks, and any Blocks the Tetromino owns is active.
    - This ownership is represented by
        a Tetromino's std::array<Block*, 4>.

- A Tetromino gains ownership of new blocks when it moves. At the same
  time, it relinquishes ownership of old blocks when it moves.

- When a Tetromino hits the ground, ownership is relinquished to the
  board, and all Blocks associated with that Tetromino remain active,
  until a line on the board is clear.

This model makes rendering (theoretically) incredibly simple. A caveat
to this is that it may make rotation calculations a bit more difficult,
as we will be relying on absolute coordinates for every block. I think
this can work, so long as we make the right abstractions in the Block
and Tetromino class going forward.